+++ src/Exceptions/PackageResolverException.php
@@ -7,4 +7,18 @@ namespace Vaimo\ComposerChangelogs\Exceptions;
 
 class PackageResolverException extends \Exception
 {
+    /**
+     * @var mixed
+     */
+    private $extraInfo;
+
+    public function setExtraInfo($value)
+    {
+        $this->extraInfo = $value;
+    }
+
+    public function getExtraInfo()
+    {
+        return $this->extraInfo;
+    }
 }