+++ src/Resolvers/VersionResolver.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Copyright © Vaimo Group. All rights reserved.
+ * See LICENSE_VAIMO.txt for license details.
+ */
+namespace Vaimo\ComposerChangelogs\Resolvers;
+
+class VersionResolver
+{
+    /**
+     * @var \Vaimo\ComposerChangelogs\Validators\ConstraintValidator 
+     */
+    private $constraintValidator;
+    
+    public function __construct()
+    {
+        $this->constraintValidator = new \Vaimo\ComposerChangelogs\Validators\ConstraintValidator();
+    }
+    
+    public function resolveValidVersion($value)
+    {
+        $segments = explode('.', $value);
+        
+        while (!$this->constraintValidator->isConstraint(implode('.', $segments)) && $segments) {
+            array_shift($segments);
+        }
+
+        return implode('.', $segments);
+    }
+}